+++ lib/src/core/vlocations.dart
@@ -5,8 +5,23 @@ import 'package:vrouter/src/wrappers/browser_helpers/browser_helpers.dart';
 import 'package:vrouter/src/wrappers/platform/platform.dart';
 
 class VLocations {
-  int _serialCount =
-      (Platform.isWeb) ? (BrowserHelpers.getHistorySerialCount() ?? 0) : 0;
+  final VRouterModes vRouterMode;
+
+  VLocations({required this.vRouterMode})
+      : _serialCount = (Platform.isWeb) ? (BrowserHelpers.getHistorySerialCount() ?? 0) : 0,
+        _locations = List<VRouteInformation?>.filled(
+                ((Platform.isWeb) ? (BrowserHelpers.getHistorySerialCount() ?? 0) : 0), null) +
+            [
+              (Platform.isWeb)
+                  ? VRouteInformation(
+                      location: BrowserHelpers.getPathAndQuery(routerMode: vRouterMode),
+                      state: Map<String, String>.from(
+                          jsonDecode((BrowserHelpers.getHistoryState() ?? '{}'))),
+                    )
+                  : null
+            ];
+
+  int _serialCount;
 
   int get serialCount => _serialCount;
 
@@ -15,21 +30,7 @@ class VLocations {
     _serialCount = newSerialCount;
   }
 
-  List<VRouteInformation?> _locations = List<VRouteInformation?>.filled(
-          ((Platform.isWeb)
-              ? (BrowserHelpers.getHistorySerialCount() ?? 0)
-              : 0),
-          null) +
-      [
-        (Platform.isWeb)
-            ? VRouteInformation(
-                location: BrowserHelpers.getPathAndQuery(
-                    routerMode: VRouterModes.history),
-                state: Map<String, String>.from(
-                    jsonDecode((BrowserHelpers.getHistoryState() ?? '{}'))),
-              )
-            : null
-      ];
+  List<VRouteInformation?> _locations;
 
   void _addLocation(VRouteInformation routeInformation) {
     _locations = _locations.sublist(0, serialCount + 1) + [routeInformation];