+++ lib/src/core/vrouter_delegate.dart
@@ -4,7 +4,6 @@ import 'dart:developer';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:path_to_regexp/path_to_regexp.dart';
-import 'package:url_strategy/url_strategy.dart';
 import 'package:vrouter/src/core/errors.dart';
 import 'package:vrouter/src/core/vlocations.dart';
 import 'package:vrouter/src/core/vnavigator_observer.dart';
@@ -14,7 +13,6 @@ import 'package:vrouter/src/core/root_vrouter.dart';
 import 'package:vrouter/src/core/route.dart';
 import 'package:vrouter/src/core/vroute_element.dart';
 import 'package:vrouter/src/core/vrouter_data.dart';
-import 'package:vrouter/src/core/vrouter_modes.dart';
 import 'package:vrouter/src/core/vrouter_scope.dart';
 import 'package:vrouter/src/helpers/empty_page.dart';
 import 'package:vrouter/src/vrouter_vroute_elements.dart';
@@ -23,8 +21,7 @@ import 'package:vrouter/src/wrappers/move_to_background.dart';
 import 'package:vrouter/src/wrappers/platform/platform.dart';
 import 'package:vrouter/src/wrappers/browser_helpers/browser_helpers.dart';
 
-class VRouterDelegate extends RouterDelegate<RouteInformation>
-    with ChangeNotifier {
+class VRouterDelegate extends RouterDelegate<RouteInformation> with ChangeNotifier {
   /// This list holds every possible routes of your app
   final List<VRouteElement> routes;
 
@@ -36,21 +33,15 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
   /// Note that if this is not implemented, every route which does not implement
   /// its own buildTransition will be given a default transition: this of a
   /// [MaterialPage] or a [CupertinoPage] depending on the platform
-  final Widget Function(Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child)? buildTransition;
+  final Widget Function(
+          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+      buildTransition;
 
   /// The duration of [VRouter.buildTransition]
   final Duration? transitionDuration;
 
   /// The reverse duration of [VRouter.buildTransition]
   final Duration? reverseTransitionDuration;
-
-  /// Two router mode are possible:
-  ///    - "hash": This is the default, the url will be serverAddress/#/localUrl
-  ///    - "history": This will display the url in the way we are used to, without
-  ///       the #. However note that you will need to configure your server to make this work.
-  ///       Follow the instructions here: [https://router.vuejs.org/guide/essentials/history-mode.html#example-server-configurations]
-  final VRouterModes mode;
 
   /// This allows you to change the initial url
   ///
@@ -90,8 +81,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     required this.routes,
     this.builder,
     this.navigatorObservers = const [],
-    Future<void> Function(VRedirector vRedirector) beforeEnter =
-        VoidVGuard.voidBeforeEnter,
+    Future<void> Function(VRedirector vRedirector) beforeEnter = VoidVGuard.voidBeforeEnter,
     Future<void> Function(
       VRedirector vRedirector,
       void Function(Map<String, String> historyState) saveHistoryState,
@@ -99,14 +89,12 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
         beforeLeave = VoidVGuard.voidBeforeLeave,
     void Function(BuildContext context, String? from, String to) afterEnter =
         VoidVGuard.voidAfterEnter,
-    Future<void> Function(VRedirector vRedirector) onPop =
-        VoidVPopHandler.voidOnPop,
+    Future<void> Function(VRedirector vRedirector) onPop = VoidVPopHandler.voidOnPop,
     Future<void> Function(VRedirector vRedirector) onSystemPop =
         VoidVPopHandler.voidOnSystemPop,
     this.buildTransition,
     this.transitionDuration,
     this.reverseTransitionDuration,
-    this.mode = VRouterModes.hash,
     this.initialUrl = '/',
   })  : _navigatorKey = GlobalKey<NavigatorState>(),
         _vNavigatorObserver = VNavigatorObserver(),
@@ -118,10 +106,29 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
           onPop: onPop,
           onSystemPop: onSystemPop,
         ) {
-    // Setup the url strategy (if hash, do nothing since it is the default)
-    if (mode == VRouterModes.history) {
-      setPathUrlStrategy();
-    }
+    WidgetsFlutterBinding.ensureInitialized().addPostFrameCallback((timeStamp) {
+      final vLocations = VRouterScope.of(_rootVRouterContext).vLocations;
+
+      // When the app starts, get the serialCount. Default to 0.
+      _serialCount = vLocations.serialCount;
+
+      // Check if this is the first route
+      if (_serialCount == 0) {
+        return pushReplacement(initialUrl);
+      } else {
+        // This happens when VRouter is rebuilt, either because:
+        //   - The entire app has been rebuilt
+        //   - VRouter key changed
+        // In this case we use _vLocations to get the current location
+        final currentLocation =
+            (vLocations.currentLocation.location.startsWith('/') ? '' : '/ ') +
+                vLocations.currentLocation.location;
+        pushReplacement(
+          currentLocation,
+          historyState: vLocations.currentLocation.state,
+        );
+      }
+    });
 
     // If we are on the web, we listen to any unload event.
     // This allows us to call beforeLeave when the browser or the tab
@@ -210,13 +217,13 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     this.queryParameters = queryParameters;
 
     // Update _vWidgetGuardMessagesRoot by removing the no-longer actives VWidgetGuards
-    for (var deactivatedVWidgetGuardMessageRoot
-        in deactivatedVWidgetGuardsMessagesRoot)
+    for (var deactivatedVWidgetGuardMessageRoot in deactivatedVWidgetGuardsMessagesRoot)
       _vWidgetGuardMessagesRoot.remove(deactivatedVWidgetGuardMessageRoot);
 
     // Update VLocations
-    VRouterScope.of(_rootVRouterContext).vLocations.setLocationAt(
-        _serialCount, VRouteInformation(location: newUrl, state: historyState));
+    VRouterScope.of(_rootVRouterContext)
+        .vLocations
+        .setLocationAt(_serialCount, VRouteInformation(location: newUrl, state: historyState));
   }
 
   /// See [VRouterMethodsHolder.pushNamed]
@@ -228,8 +235,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     bool isReplacement = false,
   }) {
     // Encode the path parameters
-    pathParameters = pathParameters
-        .map((key, value) => MapEntry(key, Uri.encodeComponent(value)));
+    pathParameters =
+        pathParameters.map((key, value) => MapEntry(key, Uri.encodeComponent(value)));
 
     // We use VRouteElement.getPathFromName
     final getPathFromNameResult = _rootVRouter.getPathFromName(
@@ -254,8 +261,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     newPath = pathToFunction(newPath)(encodedPathParameters);
 
     // Update the url with the found and completed path
-    _updateUrl(newPath,
-        queryParameters: queryParameters, isReplacement: isReplacement);
+    _updateUrl(newPath, queryParameters: queryParameters, isReplacement: isReplacement);
   }
 
   /// This should be the only way to change a url.
@@ -319,8 +325,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     late final List<VRouteElement> deactivatedVRouteElements;
     late final List<VRouteElement> reusedVRouteElements;
     late final List<VRouteElement> initializedVRouteElements;
-    late final List<VWidgetGuardMessageRoot>
-        deactivatedVWidgetGuardsMessagesRoot;
+    late final List<VWidgetGuardMessageRoot> deactivatedVWidgetGuardsMessagesRoot;
     late final List<VWidgetGuardMessageRoot> reusedVWidgetGuardsMessagesRoot;
     VRoute? newVRoute;
     if (isUrlExternal) {
@@ -354,8 +359,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       }
 
       // This copy is necessary in order not to modify newVRoute.vRouteElements
-      final newVRouteElements =
-          List<VRouteElement>.from(newVRoute.vRouteElements);
+      final newVRouteElements = List<VRouteElement>.from(newVRoute.vRouteElements);
 
       deactivatedVRouteElements = <VRouteElement>[];
       reusedVRouteElements = <VRouteElement>[];
@@ -375,8 +379,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       initializedVRouteElements = newVRouteElements
           .where(
             (newVRouteElement) =>
-                _vRoute.vRouteElements.indexWhere(
-                    (vRouteElement) => vRouteElement == newVRouteElement) ==
+                _vRoute.vRouteElements
+                    .indexWhere((vRouteElement) => vRouteElement == newVRouteElement) ==
                 -1,
           )
           .toList();
@@ -390,8 +394,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
           .toList();
       reusedVWidgetGuardsMessagesRoot = _vWidgetGuardMessagesRoot
           .where(
-            (vWidgetGuardMessageRoot) => reusedVRouteElements
-                .contains(vWidgetGuardMessageRoot.associatedVRouteElement),
+            (vWidgetGuardMessageRoot) =>
+                reusedVRouteElements.contains(vWidgetGuardMessageRoot.associatedVRouteElement),
           )
           .toList();
     }
@@ -428,10 +432,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
     if (url != null) {
       ///   1. Call beforeLeave in all deactivated [VWidgetGuard]
-      for (var vWidgetGuardMessageRoot
-          in deactivatedVWidgetGuardsMessagesRoot) {
-        await vWidgetGuardMessageRoot.vWidgetGuard
-            .beforeLeave(vRedirector, saveHistoryState);
+      for (var vWidgetGuardMessageRoot in deactivatedVWidgetGuardsMessagesRoot) {
+        await vWidgetGuardMessageRoot.vWidgetGuard.beforeLeave(vRedirector, saveHistoryState);
         if (!vRedirector.shouldUpdate) {
           await _abortUpdateUrl(
             fromBrowser: fromBrowser,
@@ -441,8 +443,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
           vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode
                   .getChildVRouteElementNode(
-                      vRouteElement:
-                          vWidgetGuardMessageRoot.associatedVRouteElement) ??
+                      vRouteElement: vWidgetGuardMessageRoot.associatedVRouteElement) ??
               _vRoute.vRouteElementNode);
           return;
         }
@@ -518,8 +519,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
           vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode
                   .getChildVRouteElementNode(
-                      vRouteElement:
-                          vWidgetGuardMessageRoot.associatedVRouteElement) ??
+                      vRouteElement: vWidgetGuardMessageRoot.associatedVRouteElement) ??
               _vRoute.vRouteElementNode);
           return;
         }
@@ -573,8 +573,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
         if (Platform.isWeb && fromBrowser && oldSerialCount != newSerialCount) {
           BrowserHelpers.browserGo(newSerialCount! - oldSerialCount);
-          await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
-              BrowserHelpers.getHistorySerialCount() == newSerialCount);
+          await BrowserHelpers.onBrowserPopState.firstWhere(
+              (element) => BrowserHelpers.getHistorySerialCount() == newSerialCount);
           _ignoreNextBrowserCalls = false;
         }
       }
@@ -591,14 +591,12 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
     /// Remove any Navigator 1.0 push
     final navigator1PushCount = _vNavigatorObserver.navigator1PushCount;
-    for (var i = 0; i < navigator1PushCount; i++)
-      _navigatorKey.currentState!.pop();
+    for (var i = 0; i < navigator1PushCount; i++) _navigatorKey.currentState!.pop();
 
     /// Leave if the url is external
     if (isUrlExternal) {
       _ignoreNextBrowserCalls = true;
-      await BrowserHelpers.pushExternal(newUri.toString(),
-          openNewTab: openNewTab);
+      await BrowserHelpers.pushExternal(newUri.toString(), openNewTab: openNewTab);
       return;
     }
 
@@ -608,13 +606,17 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     if (isReplacement) {
       _doReportBackUrlToBrowser = false;
       _ignoreNextBrowserCalls = true;
-      if (BrowserHelpers.getPathAndQuery(routerMode: mode) !=
+      if (BrowserHelpers.getPathAndQuery(
+              routerMode: VRouterScope.of(_rootVRouterContext).vRouterMode) !=
           newUri.toString()) {
-        BrowserHelpers.pushReplacement(newUri.toString(), routerMode: mode);
-        if (BrowserHelpers.getPathAndQuery(routerMode: mode) !=
+        BrowserHelpers.pushReplacement(newUri.toString(),
+            routerMode: VRouterScope.of(_rootVRouterContext).vRouterMode);
+        if (BrowserHelpers.getPathAndQuery(
+                routerMode: VRouterScope.of(_rootVRouterContext).vRouterMode) !=
             newUri.toString()) {
           await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
-              BrowserHelpers.getPathAndQuery(routerMode: mode) ==
+              BrowserHelpers.getPathAndQuery(
+                  routerMode: VRouterScope.of(_rootVRouterContext).vRouterMode) ==
               newUri.toString());
         }
       }
@@ -642,8 +644,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       newUri.toString(),
       historyState: newHistoryState,
       queryParameters: queryParameters,
-      deactivatedVWidgetGuardsMessagesRoot:
-          deactivatedVWidgetGuardsMessagesRoot,
+      deactivatedVWidgetGuardsMessagesRoot: deactivatedVWidgetGuardsMessagesRoot,
     );
     notifyListeners();
 
@@ -735,18 +736,16 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
     final List<VWidgetGuardMessageRoot> poppedVWidgetGuardsMessagesRoot =
         _vWidgetGuardMessagesRoot
-            .where((vWidgetGuardMessageRoot) => poppedVRouteElements
-                .contains(vWidgetGuardMessageRoot.associatedVRouteElement))
+            .where((vWidgetGuardMessageRoot) =>
+                poppedVRouteElements.contains(vWidgetGuardMessageRoot.associatedVRouteElement))
             .toList();
 
     /// 1. Call onPop in all popped [VWidgetGuards]
     for (var vWidgetGuardMessageRoot in poppedVWidgetGuardsMessagesRoot) {
       await vWidgetGuardMessageRoot.vWidgetGuard.onPop(vRedirector);
       if (!vRedirector.shouldUpdate) {
-        vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode
-                .getChildVRouteElementNode(
-                    vRouteElement:
-                        vWidgetGuardMessageRoot.associatedVRouteElement) ??
+        vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode.getChildVRouteElementNode(
+                vRouteElement: vWidgetGuardMessageRoot.associatedVRouteElement) ??
             _vRoute.vRouteElementNode);
         return;
       }
@@ -766,9 +765,9 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     /// 3. Call onPop of VRouter
     await _rootVRouter.onPop(vRedirector);
     if (!vRedirector.shouldUpdate) {
-      vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode
-              .getChildVRouteElementNode(vRouteElement: _rootVRouter) ??
-          _vRoute.vRouteElementNode);
+      vRedirector.redirectFunction?.call(
+          _vRoute.vRouteElementNode.getChildVRouteElementNode(vRouteElement: _rootVRouter) ??
+              _vRoute.vRouteElementNode);
       return;
     }
 
@@ -801,8 +800,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
     final List<VWidgetGuardMessageRoot> poppedVWidgetGuardsMessagesRoot =
         _vWidgetGuardMessagesRoot
-            .where((vWidgetGuardMessageRoot) => poppedVRouteElements
-                .contains(vWidgetGuardMessageRoot.associatedVRouteElement))
+            .where((vWidgetGuardMessageRoot) =>
+                poppedVRouteElements.contains(vWidgetGuardMessageRoot.associatedVRouteElement))
             .toList();
 
     /// 1. Call onSystemPop in all popping [VWidgetGuards] if implemented, else onPop
@@ -814,10 +813,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
         await vWidgetGuardMessageRoot.vWidgetGuard.onPop(vRedirector);
       }
       if (!vRedirector.shouldUpdate) {
-        vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode
-                .getChildVRouteElementNode(
-                    vRouteElement:
-                        vWidgetGuardMessageRoot.associatedVRouteElement) ??
+        vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode.getChildVRouteElementNode(
+                vRouteElement: vWidgetGuardMessageRoot.associatedVRouteElement) ??
             _vRoute.vRouteElementNode);
         return;
       }
@@ -845,9 +842,9 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       await _rootVRouter.onPop(vRedirector);
     }
     if (!vRedirector.shouldUpdate) {
-      vRedirector.redirectFunction?.call(_vRoute.vRouteElementNode
-              .getChildVRouteElementNode(vRouteElement: _rootVRouter) ??
-          _vRoute.vRouteElementNode);
+      vRedirector.redirectFunction?.call(
+          _vRoute.vRouteElementNode.getChildVRouteElementNode(vRouteElement: _rootVRouter) ??
+              _vRoute.vRouteElementNode);
       return;
     }
 
@@ -878,8 +875,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
   DefaultPopResult _defaultPop(VPopData vPopData) {
     assert(url != null);
     // Encode the path parameters
-    final pathParameters = vPopData.pathParameters
-        .map((key, value) => MapEntry(key, Uri.encodeComponent(value)));
+    final pathParameters =
+        vPopData.pathParameters.map((key, value) => MapEntry(key, Uri.encodeComponent(value)));
 
     // We don't use widget.getPathFromPop because widget.routes might have changed with a setState
     final popResult = _vRoute.vRouteElementNode.vRouteElement.getPathFromPop(
@@ -907,9 +904,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       // Integrate the given query parameters
       newUrl = Uri.tryParse(newPath)
           ?.replace(
-              queryParameters: (vPopData.queryParameters.isNotEmpty)
-                  ? vPopData.queryParameters
-                  : null)
+              queryParameters:
+                  (vPopData.queryParameters.isNotEmpty) ? vPopData.queryParameters : null)
           .toString();
 
       newVRouterData = RootVRouterData(
@@ -926,8 +922,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       newVRouterData = null;
     }
 
-    final vNodeToPop = _vRoute.vRouteElementNode
-        .getVRouteElementNodeFromVRouteElement(vPopData.elementToPop);
+    final vNodeToPop =
+        _vRoute.vRouteElementNode.getVRouteElementNodeFromVRouteElement(vPopData.elementToPop);
 
     assert(vNodeToPop != null);
 
@@ -940,16 +936,14 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     // in the nestedRoutes or stackedRoutes of [elementToPop] [VRouteElementNode]
     // We take the reversed because we when to call onPop in the deepest nested
     // [VRouteElement] first
-    final poppedVRouteElementsFromVNode =
-        vNodeToPop!.getVRouteElements().reversed.toList();
+    final poppedVRouteElementsFromVNode = vNodeToPop!.getVRouteElements().reversed.toList();
 
     // This is the list of every [VRouteElement] which should pop
     final poppedVRouteElements =
         poppedVRouteElementsFromVNode + poppedVRouteElementsFromPopResult;
 
     // elementToPop should have a duplicate so we remove it
-    poppedVRouteElements
-        .removeAt(poppedVRouteElements.indexOf(vPopData.elementToPop));
+    poppedVRouteElements.removeAt(poppedVRouteElements.indexOf(vPopData.elementToPop));
 
     return DefaultPopResult(
       vRedirector: VRedirector(
@@ -973,8 +967,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
   /// This replaces the current history state of [VRouter] with given one
   void replaceHistoryState(Map<String, String> newHistoryState) {
-    pushReplacement((url != null) ? Uri.parse(url!).path : '/',
-        historyState: newHistoryState);
+    pushReplacement((url != null) ? Uri.parse(url!).path : '/', historyState: newHistoryState);
   }
 
   /// WEB ONLY
@@ -1008,8 +1001,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
     ///   1. Call beforeLeave in all deactivated [VWidgetGuard]
     for (var vWidgetGuardMessageRoot in _vWidgetGuardMessagesRoot) {
-      await vWidgetGuardMessageRoot.vWidgetGuard
-          .beforeLeave(vRedirector, saveHistoryState);
+      await vWidgetGuardMessageRoot.vWidgetGuard.beforeLeave(vRedirector, saveHistoryState);
     }
 
     ///   2. Call beforeLeave in all deactivated [VRouteElement] and [VRouter]
@@ -1161,8 +1153,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
   }) {
     // If not on the web, this is the same as push
     if (!Platform.isWeb) {
-      return push(newUrl,
-          queryParameters: queryParameters, historyState: historyState);
+      return push(newUrl, queryParameters: queryParameters, historyState: historyState);
     }
 
     if (!newUrl.startsWith('/')) {
@@ -1197,12 +1188,10 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     Map<String, String> historyState = const {},
   }) {
     // Forming the new url by encoding each segment and placing "/" between them
-    final newUrl =
-        segments.map((segment) => Uri.encodeComponent(segment)).join('/');
+    final newUrl = segments.map((segment) => Uri.encodeComponent(segment)).join('/');
 
     // Calling push with this newly formed url
-    return push('/$newUrl',
-        queryParameters: queryParameters, historyState: historyState);
+    return push('/$newUrl', queryParameters: queryParameters, historyState: historyState);
   }
 
   /// Replace the url given a [VRouteElement] name
@@ -1250,26 +1239,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
 
   @override
   SynchronousFuture<void> setInitialRoutePath(RouteInformation configuration) {
-    final vLocations = VRouterScope.of(_rootVRouterContext).vLocations;
-
-    // When the app starts, get the serialCount. Default to 0.
-    _serialCount = vLocations.serialCount;
-
-    // Check if this is the first route
-    if (_serialCount == 0) {
-      return SynchronousFuture(pushReplacement(initialUrl));
-    } else {
-      // This happens when VRouter is rebuilt, either because:
-      //   - The entire app has been rebuilt
-      //   - VRouter key changed
-      // In this case we use _vLocations to get the current location
-      return SynchronousFuture(
-        pushReplacement(
-          vLocations.currentLocation.location,
-          historyState: vLocations.currentLocation.state,
-        ),
-      );
-    }
+    return SynchronousFuture(null);
   }
 
   /// Navigation state to app state
@@ -1278,9 +1248,8 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     if (routeInformation.location != null && !_ignoreNextBrowserCalls) {
       // Get the new state
       final newState = (Platform.isWeb)
-          ? Map<String, dynamic>.from(jsonDecode(
-              (routeInformation.state as String?) ??
-                  (BrowserHelpers.getHistoryState() ?? '{}')))
+          ? Map<String, dynamic>.from(jsonDecode((routeInformation.state as String?) ??
+              (BrowserHelpers.getHistoryState() ?? '{}')))
           : <String, dynamic>{};
 
       // Get the new serial count
@@ -1291,15 +1260,14 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       } on FormatException {}
 
       // Get the new history state
-      final newHistoryState = Map<String, String>.from(
-          jsonDecode(newState['historyState'] ?? '{}'));
+      final newHistoryState =
+          Map<String, String>.from(jsonDecode(newState['historyState'] ?? '{}'));
 
       // Check if this is the first route
       if (newSerialCount == null || newSerialCount == 0) {
         // If so, check is the url reported by the browser is the same as the initial url
         // We check "routeInformation.location == '/'" to enable deep linking
-        if (routeInformation.location == '/' &&
-            routeInformation.location != initialUrl) {
+        if (routeInformation.location == '/' && routeInformation.location != initialUrl) {
           return;
         }
       }
@@ -1324,7 +1292,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
     if (Platform.isWeb && _doReportBackUrlToBrowser) {
       BrowserHelpers.push(
         url!,
-        routerMode: mode,
+        routerMode: VRouterScope.of(_rootVRouterContext).vRouterMode,
         state: jsonEncode({
           'serialCount': _serialCount,
           'historyState': jsonEncode(historyState),
@@ -1352,16 +1320,15 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
       onNotification: (VWidgetGuardMessageRoot vWidgetGuardMessageRoot) {
         if (_vWidgetGuardMessagesRoot.indexWhere(
               (message) =>
-                  message.vWidgetGuard.key ==
-                      vWidgetGuardMessageRoot.vWidgetGuard.key &&
+                  message.vWidgetGuard.key == vWidgetGuardMessageRoot.vWidgetGuard.key &&
                   message.associatedVRouteElement ==
                       vWidgetGuardMessageRoot.associatedVRouteElement,
             ) ==
             -1) {
           _vWidgetGuardMessagesRoot.add(vWidgetGuardMessageRoot);
           WidgetsFlutterBinding.ensureInitialized().addPostFrameCallback((_) {
-            vWidgetGuardMessageRoot.vWidgetGuard.afterEnter(
-                vWidgetGuardMessageRoot.localContext, previousUrl, url!);
+            vWidgetGuardMessageRoot.vWidgetGuard
+                .afterEnter(vWidgetGuardMessageRoot.localContext, previousUrl, url!);
           });
         }
 
@@ -1396,8 +1363,7 @@ class VRouterDelegate extends RouterDelegate<RouteInformation>
                   vPopData = data;
                 } else {
                   vPopData = VPopData(
-                    elementToPop:
-                        _vRoute.vRouteElementNode.getVRouteElementToPop(),
+                    elementToPop: _vRoute.vRouteElementNode.getVRouteElementToPop(),
                     pathParameters: pathParameters,
                     queryParameters: {},
                     newHistoryState: {},
@@ -1442,7 +1408,7 @@ class RootVRouterData extends VRouterData {
     required this.historyState,
     required this.pathParameters,
     required this.queryParameters,
-  })   : state = state,
+  })  : state = state,
         super(
           key: key,
           child: child,
@@ -1490,16 +1456,14 @@ class RootVRouterData extends VRouterData {
   /// is removed from the widget tree
   ///
   /// This should be the default one, i.e. the one of [VRouter]
-  Duration? get defaultPageReverseTransitionDuration =>
-      state.reverseTransitionDuration;
+  Duration? get defaultPageReverseTransitionDuration => state.reverseTransitionDuration;
 
   /// A function to build the transition to or from this route
   ///
   /// This should be the default one, i.e. the one of [VRouter]git
   Widget Function(
-      Animation<double> animation,
-      Animation<double> secondaryAnimation,
-      Widget child)? get defaultPageBuildTransition => state.buildTransition;
+          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+      get defaultPageBuildTransition => state.buildTransition;
 
   /// See [VRouterState.push]
   void push(
@@ -1507,8 +1471,7 @@ class RootVRouterData extends VRouterData {
     Map<String, String> queryParameters = const {},
     Map<String, String> historyState = const {},
   }) =>
-      state.push(newUrl,
-          queryParameters: queryParameters, historyState: historyState);
+      state.push(newUrl, queryParameters: queryParameters, historyState: historyState);
 
   /// Pushes a new url based on url segments
   ///
@@ -1525,12 +1488,10 @@ class RootVRouterData extends VRouterData {
     Map<String, String> historyState = const {},
   }) {
     // Forming the new url by encoding each segment and placing "/" between them
-    final newUrl =
-        segments.map((segment) => Uri.encodeComponent(segment)).join('/');
+    final newUrl = segments.map((segment) => Uri.encodeComponent(segment)).join('/');
 
     // Calling push with this newly formed url
-    return push('/$newUrl',
-        queryParameters: queryParameters, historyState: historyState);
+    return push('/$newUrl', queryParameters: queryParameters, historyState: historyState);
   }
 
   /// See [VRouterState.pushNamed]
@@ -1608,8 +1569,7 @@ class RootVRouterData extends VRouterData {
         elementToPop: itemToPop,
         pathParameters: {
           ...pathParameters,
-          ...this
-              .pathParameters, // Include the previous path parameters when popping
+          ...this.pathParameters, // Include the previous path parameters when popping
         },
         queryParameters: queryParameters,
         newHistoryState: newHistoryState,
@@ -1629,8 +1589,7 @@ class RootVRouterData extends VRouterData {
           elementToPop: elementToPop,
           pathParameters: {
             ...pathParameters,
-            ...this
-                .pathParameters, // Include the previous path parameters when poping
+            ...this.pathParameters, // Include the previous path parameters when poping
           },
           queryParameters: queryParameters,
           newHistoryState: newHistoryState,
@@ -1642,8 +1601,7 @@ class RootVRouterData extends VRouterData {
       state.replaceHistoryState(historyState);
 
   static RootVRouterData of(BuildContext context) {
-    final rootVRouterData =
-        context.dependOnInheritedWidgetOfExactType<RootVRouterData>();
+    final rootVRouterData = context.dependOnInheritedWidgetOfExactType<RootVRouterData>();
     if (rootVRouterData == null) {
       throw FlutterError(
           'RootVRouterData.of(context) was called with a context which does not contain a VRouter.\n'