+++ lib/src/widgets/cupertino_vrouter.dart
@@ -422,7 +422,7 @@ class CupertinoVRouter extends StatefulWidget
 }
 
 class CupertinoVRouterState extends State<CupertinoVRouter> {
-  late final vRouterDelegate = VRouterDelegate(
+  late VRouterDelegate vRouterDelegate = VRouterDelegate(
     routes: widget.routes,
     builder: widget.builder,
     navigatorObservers: widget.navigatorObservers,
@@ -434,13 +434,34 @@ class CupertinoVRouterState extends State<CupertinoVRouter> {
     buildTransition: widget.buildTransition,
     transitionDuration: widget.transitionDuration,
     reverseTransitionDuration: widget.reverseTransitionDuration,
-    mode: widget.mode,
     initialUrl: widget.initialUrl,
   );
 
   @override
+  void didUpdateWidget(covariant CupertinoVRouter oldWidget) {
+    if (oldWidget.appRouterKey != widget.appRouterKey) {
+      vRouterDelegate = VRouterDelegate(
+        routes: widget.routes,
+        builder: widget.builder,
+        navigatorObservers: widget.navigatorObservers,
+        beforeEnter: widget.beforeEnter,
+        beforeLeave: widget.beforeLeave,
+        afterEnter: widget.afterEnter,
+        onPop: widget.onPop,
+        onSystemPop: widget.onSystemPop,
+        buildTransition: widget.buildTransition,
+        transitionDuration: widget.transitionDuration,
+        reverseTransitionDuration: widget.reverseTransitionDuration,
+        initialUrl: widget.initialUrl,
+      );
+    }
+    super.didUpdateWidget(oldWidget);
+  }
+
+  @override
   Widget build(BuildContext context) {
     return VRouterScope(
+      vRouterMode: widget.mode,
       child: CupertinoApp.router(
         backButtonDispatcher: VBackButtonDispatcher(),
         routeInformationParser: VRouteInformationParser(),