+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 - 🚨 : **BREAKING** Some class/attribute/method changed and will break your code. Read comment to know how to migrate.
 - 🔁 : **DEPRECIATED** Some class/attribute/method is depreciated. Read the documentation to know how to migrate.
 
+## \[1.1.4+1\] - 26/05/2021.
+
+* 🚨 : `VRouterScope` now holds a `vRouterMode` attribute and `VRouterDelegate` does not
+* 🛠️ : On the web, the url is stable on hot restart
+
 ## \[1.1.4\] - 26/05/2021.
 
 * 🚨 : `VRouterScope` should now be put at the top of the widget tree. If you are using `...App.router`, you must insert `VRouterScope` on top of `...App.router`.